requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
openai==1.12.0
yt-dlp==2023.12.30
moviepy==1.0.3
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
